+++ lib/src/router.dart
@@ -2,7 +2,7 @@
  * fluro
  * Created by Yakka
  * https://theyakka.com
- * 
+ *
  * Copyright (c) 2019 Yakka, LLC. All rights reserved.
  * See LICENSE for distribution and usage details.
  */
@@ -12,8 +12,8 @@ import 'dart:io';
 
 import 'package:fluro/fluro.dart';
 import 'package:fluro/src/common.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/cupertino.dart';
 
 class Router {
   static final appRouter = Router();
@@ -38,7 +38,8 @@ class Router {
     return _routeTree.matchRoute(path);
   }
 
-  bool pop<T extends Object>(BuildContext context, [T result]) => Navigator.pop(context, result);
+  bool pop<T extends Object>(BuildContext context, [T result]) =>
+      Navigator.pop(context, result);
 
   ///
   Future navigateTo(BuildContext context, String path,
@@ -46,11 +47,13 @@ class Router {
       bool clearStack = false,
       TransitionType transition,
       Duration transitionDuration = const Duration(milliseconds: 250),
-      RouteTransitionsBuilder transitionBuilder}) {
+      RouteTransitionsBuilder transitionBuilder,
+      Object arguments}) {
     RouteMatch routeMatch = matchRoute(context, path,
         transitionType: transition,
         transitionsBuilder: transitionBuilder,
-        transitionDuration: transitionDuration);
+        transitionDuration: transitionDuration,
+        routeSettings: RouteSettings(arguments: arguments));
     Route<dynamic> route = routeMatch.route;
     Completer completer = Completer();
     Future future = completer.future;
@@ -87,7 +90,8 @@ class Router {
       return MaterialPageRoute<Null>(
           settings: routeSettings,
           builder: (BuildContext context) {
-            return notFoundHandler.handlerFunc(context, parameters);
+            return notFoundHandler.handlerFunc(
+                context, parameters, routeSettings.arguments);
           });
     };
     return creator(RouteSettings(name: path), null);
@@ -118,7 +122,7 @@ class Router {
     Map<String, List<String>> parameters =
         match?.parameters ?? <String, List<String>>{};
     if (handler.type == HandlerType.function) {
-      handler.handlerFunc(buildContext, parameters);
+      handler.handlerFunc(buildContext, parameters, routeSettings.arguments);
       return RouteMatch(matchType: RouteMatchType.nonVisual);
     }
 
@@ -132,14 +136,16 @@ class Router {
               settings: routeSettings,
               fullscreenDialog: transition == TransitionType.nativeModal,
               builder: (BuildContext context) {
-                return handler.handlerFunc(context, parameters);
+                return handler.handlerFunc(
+                    context, parameters, routeSettings.arguments);
               });
         } else {
           return MaterialPageRoute<dynamic>(
               settings: routeSettings,
               fullscreenDialog: transition == TransitionType.nativeModal,
               builder: (BuildContext context) {
-                return handler.handlerFunc(context, parameters);
+                return handler.handlerFunc(
+                    context, parameters, routeSettings.arguments);
               });
         }
       } else if (transition == TransitionType.material ||
@@ -149,7 +155,8 @@ class Router {
             fullscreenDialog:
                 transition == TransitionType.materialFullScreenDialog,
             builder: (BuildContext context) {
-              return handler.handlerFunc(context, parameters);
+              return handler.handlerFunc(
+                  context, parameters, routeSettings.arguments);
             });
       } else if (transition == TransitionType.cupertino ||
           transition == TransitionType.cupertinoFullScreenDialog) {
@@ -158,7 +165,8 @@ class Router {
             fullscreenDialog:
                 transition == TransitionType.cupertinoFullScreenDialog,
             builder: (BuildContext context) {
-              return handler.handlerFunc(context, parameters);
+              return handler.handlerFunc(
+                  context, parameters, routeSettings.arguments);
             });
       } else {
         var routeTransitionsBuilder;
@@ -171,7 +179,8 @@ class Router {
           settings: routeSettings,
           pageBuilder: (BuildContext context, Animation<double> animation,
               Animation<double> secondaryAnimation) {
-            return handler.handlerFunc(context, parameters);
+            return handler.handlerFunc(
+                context, parameters, routeSettings.arguments);
           },
           transitionDuration: transitionDuration,
           transitionsBuilder: routeTransitionsBuilder,